refactor overlapping
