--- build.log
@@ -1,1 +1,0 @@
-[535/535] Linking target test-svcore-data-fileio
